+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Make `ChunkedIO#read`/`#readpartial` work with frozen chunk strings (@janko)
+
 ## 4.8.1 (2019-05-01)
 
 * Make `ChunkedIO#read`/`#readpartial` with length always return strings in binary encoding (@janko)